chore(LULA): split core; update LICENSE
